refactor: organize classes into appropriate packages
- Moved classes into specific packages to improve modularity and maintainability:
  - `muller.ui`: Contains classes related to user interaction.
  - `muller.storage`: Manages loading and saving tasks from/to files.
  - `muller.parser`: Handles command parsing.
  - `muller.task`: Encapsulates task-related classes, including `Task` and `TaskList`.
  - `muller.command`: Contains command execution classes such as `AddCommand`, `DeleteCommand`, `MarkCommand`, etc.

- Updated import statements and compilation paths to reflect the new package structure.

This reorganization enhances the clarity of the codebase, making it easier to navigate.
