+++ src/main/java/muller/command/MullerException.java
@@ -1,0 +1,7 @@
+package muller.command;
+
+public class MullerException extends Exception {
+  public MullerException(String message) {
+    super(message);
+  }
+}
